move milestone hook logic out into its own file
